First pass at adding api caller tests - don't commit migration
